Server is working -- insertRandomData still failing
